requirements.txt
Django>=1.11,<2.0
django-dash>=0.5.3
django-celery-results>=1.0.1
django-jsonfield>=1.0.1
django-pandas>=0.5.1
    celery>=4.0.2,<=4.1.0
    kombu==4.1.0
tornado==5.1
flower>=0.9.2
PyYAML>=3.12
pyaml>=17.12.1
psycopg2-binary>=2.7
rest-pandas>=1.0.0
matplotlib>=2.0.2
django-extensions>=2.1.0
cherrypy>=17.0.0
jupyter>=1.0.0
paramiko>=2.3.1
django-fsm>=2.6.0
xlrd>=1.1.0
